+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - :rocket: added types to global members
 - :rocket: added _I switch to 'IFrame' frame_
 - :rocket: added _I switch to 'window name or title' window_
+- :rocket: added steps to work with multiple browser contexts
 
 ## [0.0.22]
 - :beetle: fixed exports of mock and poDefine