+++ CHANGELOG.md
@@ -4,11 +4,13 @@ All notable changes to the "@qavajs/steps-playwright" will be documented in this
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.0.24]
+- :rocket: added steps to work with multiple browser contexts
+
 ## [0.0.23]
 - :rocket: added types to global members
 - :rocket: added _I switch to 'IFrame' frame_
 - :rocket: added _I switch to 'window name or title' window_
-- :rocket: added steps to work with multiple browser contexts
 
 ## [0.0.22]
 - :beetle: fixed exports of mock and poDefine